# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NexusID API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List job listings
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [full-time, part-time, contract, remote]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Jobs page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobsPage"

    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobBody"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{jobId}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a single job listing
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{jobId}/apply:
    post:
      operationId: applyToJob
      tags: [jobs]
      summary: Apply to a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobApplicationBody"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobApplication"
...
  /users/{userId}/power-score:
    get:
      operationId: getPowerScore
      tags: [users]
      summary: Get Power Score and breakdown for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Power score result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PowerScoreResult"

  # ─── Ghost Mode ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        username:
          type: string
        displayName:
          type: string
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        coverUrl:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        occupation:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
        experience:
          type: array
          items:
            $ref: "#/components/schemas/Experience"
        didDocument:
          type: object
          nullable: true
          additionalProperties: true
        followersCount:
          type: integer
        followingCount:
          type: integer
        postsCount:
          type: integer
        ghostMode:
          type: boolean
        isFollowing:
          type: boolean
        bountiesWon:
          type: integer
        createdAt:
          type: string
        usernameChangedAt:
          type: string
          nullable: true
        lastSeenAt:
          type: string
          nullable: true
          description: Last time this user pinged the server. Powers the online dot.
        powerScore:
          type: ["integer", "null"]
        powerRank:
          type: ["string", "null"]
        isAdmin:
          type: boolean
          description: True when this user is a moderator/admin.
        verificationTier:
          type: string
          nullable: true
          enum: [silver, blue, null]
          description: |
            Identity-verification badge shown next to the display name.
            "silver" = identity verified, "blue" = notable / premium account,
            null = unverified. Granted by an admin via
            PATCH /admin/users/:userId/verification.
        chatScreenshotsTaken:
          type: integer
          description: |
            Public counter of how many times this user has screenshotted a
            DM thread (mobile only). Surfaced as a warning chip on the
            profile so peers can decide whether to share sensitive content.
      required: [id, clerkId, username, displayName, skills, experience, followersCount, followingCount, postsCount, ghostMode, isFollowing, bountiesWon, createdAt]
...
    Experience:
      type: object
      properties:
        title:
          type: string
        company:
          type: string
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        current:
          type: boolean
        description:
          type: string
          nullable: true
      required: [title, company, startDate, current]
...
    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          type: string
        location:
          type: string
        type:
          type: string
          enum: [full-time, part-time, contract, remote]
        description:
          type: string
        skills:
          type: array
          items:
            type: string
        salaryMin:
          type: integer
          nullable: true
        salaryMax:
          type: integer
          nullable: true
        poster:
          $ref: "#/components/schemas/UserProfile"
        applicantsCount:
          type: integer
        aiMatchScore:
          type: number
          nullable: true
        createdAt:
          type: string
      required: [id, title, company, location, type, description, skills, poster, applicantsCount, createdAt]
    CreateJobBody:
      type: object
      properties:
        title:
          type: string
        company:
          type: string
        location:
          type: string
        type:
          type: string
          enum: [full-time, part-time, contract, remote]
        description:
          type: string
        skills:
          type: array
          items:
            type: string
        salaryMin:
          type: integer
        salaryMax:
          type: integer
      required: [title, company, location, type, description, skills]
    JobsPage:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
        hasMore:
          type: boolean
      required: [jobs, total, hasMore]
    JobApplicationBody:
      type: object
      properties:
        coverLetter:
          type: string
      required: [coverLetter]
    JobApplication:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        userId:
          type: string
        coverLetter:
          type: string
        status:
          type: string
          enum: [pending, reviewing, accepted, rejected]
        createdAt:
          type: string
      required: [id, jobId, userId, coverLetter, status, createdAt]

    # ─── Communities ───────────────────────────────────────────────────────────
...
    PowerScoreResult:
      type: object
      properties:
        score:
          type: integer
        rank:
          type: string
        breakdown:
          type: object
          properties:
            network:
              type: integer
            content:
              type: integer
            activity:
              type: integer
            reputation:
              type: integer
          required: [network, content, activity, reputation]
      required: [score, rank, breakdown]

    # ─── Ghost Mode ────────────────────────────────────────────────────────────
...